backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.1",
    "arq>=0.26.3",
    "asyncpg>=0.31.0",
    "boto3>=1.42.29",
    "fastapi>=0.128.0",
    "firebase-admin>=7.1.0",
    "google-cloud-aiplatform>=1.133.0",
    "greenlet>=3.3.0",
    "honcho>=2.0.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pip>=25.3",
    "postgresql-audit>=0.17.1",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyotp>=2.9.0",
    "python-ulid>=3.1.0",
    "secure>=1.0.1",
    "sentry-sdk[fastapi]>=2.49.0",
    "slowapi>=0.1.9",
    "spacy>=3.8.11",
    "sqladmin>=0.22.0",
    "sqlalchemy>=2.0.45",
    "sse-starlette>=3.1.2",
    "stripe>=14.1.0",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "faker>=40.1.2",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "COM", "C4", "DTZ", "PIE", "T20", "RET", "SIM", "ARG", "PTH", "ERA", "PL", "RUF"]
ignore = ["COM812", "PLR0913", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004", "RUF012", "T201"]
"migrations/versions/*" = ["E501"]
"app/core/pii_masking.py" = ["ARG001", "S110"]
"app/core/audit_sql.py" = ["E501"]
"app/core/config.py" = ["S104"]
"app/core/logging.py" = ["RUF005"]
"main.py" = ["T201"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"



